Alternative way to avoid overwrite of price
using a transparent drawable at the end of the name view, so that
space for price is reserved only on the last row of the name.

This is a proof-of-concept so far. To make it real, need to compute
the actual width of the Price TextView and set it as the drawable width,
in place of the currently hard-coded 20.